requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
jinja2
google-auth
gunicorn
sqlalchemy
PyJWT
requests
yfinance
pytz
apscheduler
python-dotenv>=1.0
kiteconnect
fastapi_utils
typing_inspect
openpyxl
pandas
pydantic